Football scouting website
The main goal of this project is to create a thorough football scouting platform to address the inherent difficulties in the football recruiting and scouting process. This platform uses machine learning algorithms, data-driven insights, and contemporary technology to transform talent identification and recruitment. The platform aims to improve the efficacy, efficiency, and transparency of the scouting process by offering a central location for football players and agents to interact, communicate, and work together.